Completely rework progress tracking, use Content-Length for progress info
Fixes #35 — that kind of error is no longer possible. The downside of this approach is the progress bar will appear to move in "chunks" as every file is given equal weight, rather than it depending on the file's size. However, this was already how it worked with Packwiz, as that format doesn't give us size info up-front like the native format does. C'est la vie.